pyproject.toml
[tool.poetry]
name = "googletools"
version = "0.1.0"
description = ""
authors = ["Tim Mees <tim.mees83@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.4.0"
google-api-core = "^2.5.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.4.6"
google-api-python-client = "^2.36.0"
xdg = "^5.1.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
flake8-bugbear = "^22.1.11"
black = "^22.1.0"
isort = "^5.10.1"
bandit = "^1.7.2"
safety = "^1.10.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
